* main: (24 commits)
  chore(deps): update dependency nuqs to v2.3.0 (#549)
  chore(deps): update dependency viem to v2.22.1 (#548)
  chore(deps): update dependency @types/node to v22.10.3 (#547)
  chore(deps): update dependency viem to v2.22.0 (#546)
  fix: e2e tests (#545)
  fix: set workspace command
  fix: add quotes
  fix: use env
  fix: use github cli to save workspace config
  chore: update package versions [skip ci]
  fix: prefer using unique names over id's (#508)
  chore(deps): update dependency ts-morph to v25 (#543)
  chore(deps): update dependency @commander-js/extra-typings to v13 (#544)
  chore(deps): update skandalik/save-secret action to v1.0.1 (#542)
  chore(deps): update dependency viem to v2.21.60 (#541)
  fix: re-use the same workspace for the E2E tests for one week (#539)
  chore(deps): update dependency viem to v2.21.58 (#538)
  chore: update package versions [skip ci]
  fix: predicable names (#537)
  chore: update package versions [skip ci]
  ...

+++ test/create-new-project.e2e.test.ts
@@ -1,7 +1,7 @@
 import { afterAll, afterEach, beforeAll, describe, expect, setDefaultTimeout, test } from "bun:test";
 import { copyFile, readFile, rmdir, stat } from "node:fs/promises";
 import { join } from "node:path";
-import { type DotEnv, loadEnv } from "@settlemint/sdk-utils";
+import { type DotEnv, exists, loadEnv } from "@settlemint/sdk-utils";
 import { $ } from "bun";
 import {
   getSubgraphYamlConfig,
@@ -49,17 +49,13 @@ describe("Setup a project using the SDK", () => {
     ).result;
     expect((await stat(projectDir)).isDirectory()).toBeTrue();
     expect(output).toInclude("Your project is ready to go!");
-    await copyFile(join(__dirname, "../.env"), join(projectDir, ".env"));
-    await copyFile(join(__dirname, "../.env.local"), join(projectDir, ".env.local"));
+    if (await exists(join(__dirname, "../.env"))) {
+      await copyFile(join(__dirname, "../.env"), join(projectDir, ".env"));
+    }
+    if (await exists(join(__dirname, "../.env.local"))) {
+      await copyFile(join(__dirname, "../.env.local"), join(projectDir, ".env.local"));
+    }
   });
-
-  // test("Create Minio storage on the platform", () => {
-  //   // Optional, can be done later
-  // });
-
-  // test("Create custom deployment on the platform", () => {
-  //   // Optional, can be done later
-  // });
 
   test("Validate that .env file has the correct values", async () => {
     const env: Partial<DotEnv> = await loadEnv(false, false, projectDir);
